added child dashboard, fixed folder
