+++ src/App.js
@@ -1,11 +1,11 @@
 import React from 'react';
-import Landing from './modules/Landing/components/Landing/Landing';
+import Landing from './Components/Landing/components/Landing';
 import { Routes, Route } from "react-router-dom";
-import Login from './views/ParentViews/Login/components/Login/Login';
+import Login from './Components/ParentViews/Login/components/Login';
 import './common/styles/GlobalStyles.css'
-import ChildLogin from './views/ChildViews/components/Login/ChildLogin';
-import Register from './views/ParentViews/Register/components/Register';
-
+import ChildLogin from './Components/ChildViews/Login/components/ChildLogin'
+import Register from './Components/ParentViews/Register/components/Register';
+import ChildDashboard from './Components/ChildViews/Dashboard/components/ChildDashboard';
 
 function App() {
 
@@ -28,6 +28,10 @@ function App() {
           <Route
             path='/child-login'
             element={<ChildLogin />} 
+          />
+          <Route
+            path='/child-dashboard'
+            element={<ChildDashboard />}
           />
       </Routes>
       </nav>
