+++ src/Components/ChildViews/Dashboard/routes/ChildDashboardPath.js
@@ -1,0 +1,9 @@
+import React from 'react'
+import ChildDashboard from '../components/ChildDashboard'
+
+export const childDashboardPath = '/child-dashboard'
+
+export const childDashboardRoute = {
+    path: childDashboardPath,
+    element: <ChildDashboard />
+}
