+++ src/Components/ChildViews/Dashboard/styles/ChildDashboard.css
@@ -1,0 +1,16 @@
+.userTitle{
+    font-size: 20px;
+    text-align: center;
+    padding-bottom: 5px;
+}
+.taskTable{
+   text-align: center;
+   background-color:#B3C8E5;
+   opacity: 80%;
+   border-radius: 15px;
+}
+.taskTitle{
+    font-size: 20px;
+}
+
+
