+++ src/Components/Landing/components/Footer.js
@@ -1,0 +1,13 @@
+import React from 'react'
+import '../styles/Footer.css'
+
+const Footer = () => {
+    return (
+        <div className='Footer'>
+            <h4>About us</h4>
+            <p>© 2023</p>
+        </div>
+    )
+}
+
+export default Footer;
