+++ src/Components/Landing/routes/LandingRoute.js
@@ -1,0 +1,9 @@
+import React from 'react'
+import Landing from '../components/Landing'
+
+export const landingPath = '/'
+
+export const landingRoute = {
+    path: landingPath,
+    element: <Landing />
+}
