+++ src/Components/Landing/styles/ChildLoginButton.css
@@ -1,0 +1,22 @@
+.ChildLoginButton{
+    display: block;
+    background: rgba(109,146,229,255);
+    border-radius: 15px;
+    border: transparent;
+    width: 100%;
+    height: 35px;
+    letter-spacing: 1px;
+}
+.ChildLoginLink{
+    color: white;
+    text-decoration: none;
+    font-size: 14px;
+    font-weight: bold;
+}
+.ChildLoginButton:hover {
+    color: black;
+}
+.ChildLoginButton:active {
+    box-shadow: 0 4px gray;
+    transform: translateY(4px);
+}
