+++ src/Components/ParentViews/Login/routes/LoginRoute.js
@@ -1,0 +1,9 @@
+import React from 'react'
+import Login from '../components/Login'
+
+export const loginPath = '/'
+
+export const loginRoute = {
+    path: loginPath,
+    element: <Login />
+}
