+++ src/Components/ParentViews/routes/ParentDashboard.js
@@ -1,0 +1,9 @@
+import React from 'react'
+import Dashboard from '../components/Dashboard/Dashboard'
+
+export const dashboardPath = '/'
+
+export const dashboardRoute = {
+    path: dashboardPath,
+    element: <Dashboard />
+}
