--- src/modules/Landing/routes/LandingRoute.js
@@ -1,9 +1,0 @@
-import React from 'react'
-import Landing from '../components/Landing/Landing'
-
-export const landingPath = '/'
-
-export const landingRoute = {
-    path: landingPath,
-    element: <Landing />
-}
