--- src/modules/Landing/styles/ChildLoginButton.css
@@ -1,15 +1,0 @@
-.ChildLoginButton{
-    display: block;
-    background: rgba(109,146,229,255);
-    border-radius: 15px;
-    border: transparent;
-    width: 100%;
-    height: 35px;
-    letter-spacing: 1px;
-}
-.ChildLoginLink{
-    color: white;
-    text-decoration: none;
-    font-size: 14px;
-    font-weight: bold;
-}
