--- src/views/ChildViews/components/Dashboard/ChildDashboard.js
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const ChildDashboard = () => {
-  return (
-    <div>ChildDashboard</div>
-  )
-}
-
-
-
-export default ChildDashboard;
