--- src/views/ChildViews/routes/ChildDashboardPath.js
@@ -1,9 +1,0 @@
-import React from 'react'
-import ChildDashboard from '../components/Dashboard/ChildDashboard'
-
-export const childDashboardPath = '/'
-
-export const childDashboardRoute = {
-    path: childDashboardPath,
-    element: <ChildDashboard />
-}
