--- src/views/ParentViews/Login/routes/LoginRoute.js
@@ -1,9 +1,0 @@
-import React from 'react'
-import Login from '../components/Login/Login'
-
-export const loginPath = '/'
-
-export const loginRoute = {
-    path: loginPath,
-    element: <Login />
-}
